+++ src/index.pug
@@ -35,7 +35,6 @@ html(lang='ja')
       #info
     #data-credits
     script(src='loadSounds.js')
-    script(src='usafuya.js')
     script(src='loading.js')
     script(src='usapyon_achievement.js')
     script(src='usapyon_ui.js')
