+++ src/usapyon_ui.js
@@ -1,7 +1,3 @@
-$("#1").click(function() {
-  usafuya();
-});
-
 $(window).keydown(function() {
   return false;
 });
